feat: 管理端工作台、导出Excel报表功能接口实现

+++ sky-server/src/main/java/com/sky/mapper/UserMapper.java
@@ -41,5 +41,5 @@ public interface UserMapper {
      * @param endTime
      * @return
      */
-    Long countByCreateTime(LocalDateTime beginTime, LocalDateTime endTime);
+    Integer countByCreateTime(LocalDateTime beginTime, LocalDateTime endTime);
 }
